\alias{ROOXML}
\name{ROOXML}
\title{the ROOXML package}
\description{
  This package provides an infrastructure
  for working with the common parts of Office Open files
  for Word, Excel, Power Point, etc.
}
\author{Duncan Temple Lang}
\seealso{
 The RWordXML and RExcelXML packages
}
\keyword{package}
\keyword{IO}
